producer/requirements.txt
kafka-python
selenium
newspaper3k
gnews
undetected_chromedriver